# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Techno Terminal API",
    "description": "RESTful backend for the Techno Terminal CRM & Operations System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/competitions": {
      "get": {
        "tags": [
          "Competitions",
          "Competitions"
        ],
        "summary": "List all competitions",
        "operationId": "list_competitions_api_v1_competitions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_CompetitionDTO__"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Competitions",
          "Competitions"
        ],
        "summary": "Create a new competition",
        "operationId": "create_competition_api_v1_competitions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompetitionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CompetitionDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/competitions/{competition_id}": {
      "get": {
        "tags": [
          "Competitions",
          "Competitions"
        ],
        "summary": "Get single competition details",
        "operationId": "get_competition_api_v1_competitions__competition_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "competition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Competition Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CompetitionDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/competitions/{competition_id}/categories": {
      "get": {
        "tags": [
          "Competitions",
          "Competitions"
        ],
        "summary": "List categories for a competition",
        "operationId": "list_categories_api_v1_competitions__competition_id__categories_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "competition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Competition Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_CompetitionCategoryDTO__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Competitions",
          "Competitions"
        ],
        "summary": "Add a category to a competition",
        "operationId": "add_category_api_v1_competitions__competition_id__categories_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "competition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Competition Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCategoryInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_CompetitionCategoryDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/competitions/register": {
      "post": {
        "tags": [
          "Competitions",
          "Competitions"
        ],
        "summary": "Register a team for a competition",
        "operationId": "register_team_api_v1_competitions_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterTeamInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TeamRegistrationResultDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/competitions/{competition_id}/categories/{category_id}/teams": {
      "get": {
        "tags": [
          "Competitions",
          "Competitions"
        ],
        "summary": "List teams in a competition category",
        "operationId": "list_teams_in_category_api_v1_competitions__competition_id__categories__category_id__teams_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "competition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Competition Id"
            }
          },
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_TeamWithMembersDTO__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/competitions/team-members/{team_member_id}/pay": {
      "post": {
        "tags": [
          "Competitions",
          "Competitions"
        ],
        "summary": "Mark competition fee as paid (bypass Finance Desk)",
        "operationId": "mark_fee_as_paid_api_v1_competitions_team_members__team_member_id__pay_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Member Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddCategoryInput": {
        "properties": {
          "competition_id": {
            "type": "integer",
            "title": "Competition Id"
          },
          "category_name": {
            "type": "string",
            "title": "Category Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "competition_id",
          "category_name"
        ],
        "title": "AddCategoryInput"
      },
...
      "ApiResponse_CompetitionCategoryDTO_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompetitionCategoryDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "ApiResponse[CompetitionCategoryDTO]"
      },
      "ApiResponse_CompetitionDTO_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompetitionDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "ApiResponse[CompetitionDTO]"
      },
...
      "ApiResponse_NoneType_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "type": "null",
            "title": "Data"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "ApiResponse[NoneType]"
      },
...
      "ApiResponse_TeamRegistrationResultDTO_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TeamRegistrationResultDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "ApiResponse[TeamRegistrationResultDTO]"
      },
...
      "ApiResponse_list_CompetitionCategoryDTO__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CompetitionCategoryDTO"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "ApiResponse[list[CompetitionCategoryDTO]]"
      },
      "ApiResponse_list_CompetitionDTO__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CompetitionDTO"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "ApiResponse[list[CompetitionDTO]]"
      },
...
      "ApiResponse_list_TeamWithMembersDTO__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TeamWithMembersDTO"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "ApiResponse[list[TeamWithMembersDTO]]"
      },
...
      "CompetitionCategoryDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "competition_id": {
            "type": "integer",
            "title": "Competition Id"
          },
          "category_name": {
            "type": "string",
            "title": "Category Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "competition_id",
          "category_name"
        ],
        "title": "CompetitionCategoryDTO"
      },
      "CompetitionDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "edition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edition"
          },
          "competition_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Competition Date"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "fee_per_student": {
            "type": "number",
            "title": "Fee Per Student",
            "default": 0.0
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CompetitionDTO"
      },
...
      "CreateCompetitionInput": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "edition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edition"
          },
          "competition_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Competition Date"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "fee_per_student": {
            "type": "number",
            "title": "Fee Per Student",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateCompetitionInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RegisterTeamInput": {
        "properties": {
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "team_name": {
            "type": "string",
            "title": "Team Name"
          },
          "student_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Student Ids"
          },
          "coach_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coach Id"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          }
        },
        "type": "object",
        "required": [
          "category_id",
          "team_name",
          "student_ids"
        ],
        "title": "RegisterTeamInput"
      },
...
      "TeamDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "team_name": {
            "type": "string",
            "title": "Team Name"
          },
          "coach_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coach Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category_id",
          "team_name"
        ],
        "title": "TeamDTO"
      },
      "TeamMemberDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "team_id": {
            "type": "integer",
            "title": "Team Id"
          },
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "member_share": {
            "type": "number",
            "title": "Member Share",
            "default": 0.0
          },
          "fee_paid": {
            "type": "boolean",
            "title": "Fee Paid",
            "default": false
          },
          "payment_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "team_id",
          "student_id"
        ],
        "title": "TeamMemberDTO"
      },
      "TeamRegistrationResultDTO": {
        "properties": {
          "team": {
            "$ref": "#/components/schemas/TeamDTO"
          },
          "members_added": {
            "type": "integer",
            "title": "Members Added"
          }
        },
        "type": "object",
        "required": [
          "team",
          "members_added"
        ],
        "title": "TeamRegistrationResultDTO",
        "description": "Returned by register_team"
      },
      "TeamWithMembersDTO": {
        "properties": {
          "team": {
            "$ref": "#/components/schemas/TeamDTO"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/TeamMemberDTO"
            },
            "type": "array",
            "title": "Members"
          }
        },
        "type": "object",
        "required": [
          "team",
          "members"
        ],
        "title": "TeamWithMembersDTO"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}